requirements.txt
# this is for Heroku and other servers.
# Locally, you should install requirements_base.txt
-r requirements_base.txt
psycopg2>=2.5.1
oTree==3.2.5
Django==2.2.12
django-model-utils
